feat!: adds more integration test adjustments
BREAKING CHANGE: The app is still in a broken state, as I'm going to
pivot now to business tests.
I think there are two types of tests; business tests, and UI tests. I'm
going to start with the business tests now, as they are a bit more
important than the UI tests. I can test them directly, the business
tests are a bit more complex and delicate.